refactor(lsp): consolidate incumbent wait policy
